Gjorde om animation klassen
